build!: golangci-lintを使用し、コードフォーマット/リンターをコミット時に行う
- .gitignore更新
- node.js v26更新
- miseのminimum release ageを7日に設定

- record/padding sizeチェック追加

+++ contentencoding_test.go
@@ -14,6 +14,6 @@ import (
 )
 
 func TestContentEncoding_Padding(t *testing.T) {
-	assert.Equal(t, uint32(2), AESGCM.Padding())
-	assert.Equal(t, uint32(1), AES128GCM.Padding())
+	assert.Equal(t, 2, AESGCM.Padding())
+	assert.Equal(t, 1, AES128GCM.Padding())
 }
